requirements.txt
aniso8601==9.0.1
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.0
Flask==2.1.0
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
google-api-core==2.7.3
google-api-python-client==2.46.0
google-auth==2.6.6
google-auth-httplib2==0.1.0
googleapis-common-protos==1.56.0
gunicorn==20.1.0
httplib2==0.20.4
idna==3.3
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
joblib==1.1.0
MarkupSafe==2.1.1
numpy==1.22.3
oauth2client==4.1.3
pandas==1.4.2
protobuf==3.20.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyDrive==1.3.1
pyparsing==3.0.8
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
requests==2.27.1
rsa==4.8
scikit-learn==1.1.0
scipy==1.8.0
six==1.16.0
sklearn==0.0
threadpoolctl==3.1.0
uritemplate==4.1.1
urllib3==1.26.9
Werkzeug==2.1.0
zipp==3.7.0
